requirements.txt
ipykernel
langchain
python-dotenv
langchain-community
Pypdf
bs4
arxiv
pymupdf
wikipedia
langchain-text-splitters
chromadb
sentence_transformers
langchain_huggingface
faiss-cpu
langchain_chroma
langchain-google-genai
google-generativeai
streamlit
langchain_groq
langchain_core
fastapi
uvicorn
langserve
bs4
arxiv 
wikipedia
duckduckgo-search